Introduce wrappers and cleanup code

+++ src/kernel/plic.rs
@@ -8,13 +8,13 @@ pub fn initialize() {
 }
 
 pub fn configure_cpu() {
-    let context = context(riscv64::read_tp() as u32);
+    let context = context(riscv64::read_tp());
     plic::enable_interrupt(context, VIRT_UART0_SOURCE_NUMBER);
     plic::set_priority_threshold(context, 0);
 }
 
 // Learned from the xv6-riscv repository.
 // This setting may exist in the hw/riscv/virt.c of qemu repository.
-fn context(hart_id: u32) -> u32 {
-    1 + hart_id * 2
+fn context(hart_id: u64) -> u32 {
+    1 + (hart_id as u32) * 2
 }
